+++ src/Dumper/Config/Validation/QueryValidator.php
@@ -16,12 +16,13 @@ final class QueryValidator
      * @var string[]
      */
     private array $statements = [
-        'alter', 'analyse', 'backup', 'binlog', 'cache', 'change', 'close', 'commit', 'create',
-        'deallocate', 'declare', 'delete', 'describe', 'drop', 'execute', 'explain', 'fetch', 'flush',
-        'get', 'grant', 'help', 'install', 'kill', 'load', 'lock', 'open', 'optimize', 'prepare',
-        'purge', 'rename', 'repair', 'reset', 'resignal', 'revoke', 'rollback', 'savepoint', 'select',
-        'set', 'password', 'show', 'shutdown', 'signal', 'start', 'truncate', 'uninstall', 'unlock',
-        'update', 'use', 'xa',
+        'alter', 'analyze', 'backup', 'begin', 'binlog', 'cache', 'call', 'change', 'check', 'checksum',
+        'close', 'commit', 'create', 'deallocate', 'declare', 'delete', 'describe', 'do', 'drop', 'end',
+        'execute', 'explain', 'fetch', 'flush', 'get', 'grant', 'help', 'insert', 'import', 'install',
+        'kill', 'load', 'lock', 'open', 'optimize', 'prepare', 'purge', 'rename', 'repair', 'replace',
+        'reset', 'resignal', 'restart', 'revoke', 'rollback', 'savepoint', 'select', 'set', 'password',
+        'show', 'shutdown', 'signal', 'start', 'stop', 'truncate', 'uninstall', 'unlock', 'update',
+        'use', 'xa',
     ];
 
     /**